requirements.txt
azure-ai-projects==1.0.0
aiohttp==3.12.15
azure-identity==1.24.0
python-dotenv==1.1.1
azure-monitor-opentelemetry
opentelemetry-sdk